Move another function to gui utils
Base utils should only use the base dependencies
