+++ damnit/util.py
@@ -1,12 +1,9 @@
-import glob
 import sys
 from datetime import datetime, timezone
 
 import numpy as np
 import pandas as pd
 from pandas.api.types import infer_dtype
-
-from .context import add_to_h5_file
 
 
 def timestamp2str(timestamp):
@@ -32,18 +29,6 @@ def bool_to_numeric(data):
 
 def fix_data_for_plotting(data):
     return bool_to_numeric(make_finite(data))
-
-
-def delete_variable(db, name):
-    # Remove from the database
-    db.delete_variable(name)
-
-    # And the HDF5 files
-    for h5_path in glob.glob(f"{db.path.parent}/extracted_data/*.h5"):
-        with add_to_h5_file(h5_path) as f:
-            if name in f:
-                del f[f".reduced/{name}"]
-                del f[name]
 
 
 def isinstance_no_import(obj, mod: str, cls: str):
