+++ pyproject.toml
@@ -17,6 +17,7 @@ readme = "README.md"
 dependencies = [
     "h5netcdf>=1.4.1",
     "h5py",
+    "numpy",
     "orjson",  # used in plotly for faster json serialization
     "pandas",
     "plotly",
